speedups to R_WorldNode, and some shrinkage on the surface struct
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@78 d7cf8633-e32d-0410-b094-e92efae38249

+++ gl_rmain.c
@@ -63,7 +63,7 @@ mleaf_t		*r_viewleaf, *r_oldviewleaf;
 
 texture_t	*r_notexture_mip;
 
-int		d_lightstylevalue[256];	// 8.8 fraction of base light value
+unsigned short	d_lightstylevalue[256];	// 8.8 fraction of base light value
 
 
 void R_MarkLeaves (void);
